Before:
```
BenchmarkSandSlabsPart1-10         1      3286466875 ns/op      182566344 B/op     17827 allocs/op
BenchmarkSandSlabsPart2-10         1      3277207750 ns/op      182564072 B/op     17821 allocs/op
```

After:
```
BenchmarkSandSlabsPart1-10        61        19153398 ns/op      82040177 B/op      17271 allocs/op
BenchmarkSandSlabsPart2-10        60        19138804 ns/op      82040241 B/op      17271 allocs/op
```